Apply files structure rule with one type in one file to EventData.
