+++ Sources/EventStoreDB/Event/ContentType.swift
@@ -1,0 +1,14 @@
+//
+//  ContentType.swift
+//
+//
+//  Created by 卓俊諺 on 2024/6/2.
+//
+
+import Foundation
+
+public enum ContentType: String, Codable, Sendable {
+    case unknown
+    case json = "application/json"
+    case binary = "application/octet-stream"
+}
