+++ Sources/EventStoreDB/Event/EventStoreEvent.swift
@@ -1,0 +1,14 @@
+//
+//  EventStoreEvent.swift
+//
+//
+//  Created by 卓俊諺 on 2024/6/2.
+//
+
+import Foundation
+
+public protocol EventStoreEvent {
+    var id: UUID { get }
+    var eventType: String { get }
+    var contentType: ContentType { get }
+}
